+++ app/form/administration/users_directory/settings.js
@@ -9,17 +9,20 @@
  * @copyright     Copyright (c) Passbolt SARL (https://www.passbolt.com)
  * @license       https://opensource.org/licenses/AGPL-3.0 AGPL License
  * @link          https://www.passbolt.com Passbolt(tm)
- * @since         2.5.0
+ * @since         2.6.0
  */
 import FeedbackComponent from 'passbolt-mad/form/feedback';
 import Form from 'passbolt-mad/form/form';
 import RadioComponent from 'passbolt-mad/form/element/radio';
+import domEvents from 'can-dom-events';
 import DropdownComponent from 'passbolt-mad/form/element/dropdown';
 import TextboxComponent from 'passbolt-mad/form/element/textbox';
 import ToggleButtonComponent from 'passbolt-mad/form/element/toggle_button';
 import User from 'app/model/map/user';
+import 'chosen-js/chosen.jquery';
 
 import template from 'app/view/template/form/administration/users_directory/settings.stache!';
+import { isFunction } from 'util';
 
 const UsersDirectorySettingsForm = Form.extend('passbolt.form.administration.users_directory.Settings', /** @static */ {
 
@@ -41,7 +44,8 @@ const UsersDirectorySettingsForm = Form.extend('passbolt.form.administration.use
    */
   loadAndStart: function(usersDirectorySettings) {
     this.options.usersDirectorySettings = usersDirectorySettings;
-    this.start();
+    this._retrieveUsers()
+      .then(() => this.start());
   },
 
   /**
@@ -49,7 +53,6 @@ const UsersDirectorySettingsForm = Form.extend('passbolt.form.administration.use
    */
   afterStart: function() {
     this._initForm(this.options.usersDirectorySettings);
-    this._initEditForm();
     this.state.loaded = true;
     this.state.hidden = false;
     this._super();
@@ -64,28 +67,70 @@ const UsersDirectorySettingsForm = Form.extend('passbolt.form.administration.use
   },
 
   /**
+   * Retrieve passbolt users.
+   */
+  _retrieveUsers: function() {
+    const findOptions = {
+      contain: {profile: 1}
+    };
+    return User.findAll(findOptions)
+      .then(users => {
+        this._users = users.reduce((carry, user) => {
+          carry[user.id] = `${user.profile.fullName()} (${user.username})`;
+          return carry;
+        }, {});
+        this._admins = users.filter(user => user.role.name == 'admin').reduce((carry, user) => {
+          carry[user.id] = `${user.profile.fullName()} (${user.username})`;
+          return carry;
+        }, {});
+      });
+  },
+
+  /**
    * Init the form.
    */
-  _initForm: function(usersDirectorySettings) {
-    this._initFormGeneralSwitch(usersDirectorySettings);
+  _initForm: function() {
+    const settings = this.options.usersDirectorySettings;
+    this._initFormGeneralSwitch();
     this._initFormCredantialsSection();
     this._initFormDirectorySection();
     this._initFormSynchronisationSection();
-    this.load({'UsersDirectorySettings': usersDirectorySettings});
+    this.load({'UsersDirectorySettings': settings});
+    this._afterLoad();
+  },
+
+  /**
+   * After the form is loaded
+   */
+  _afterLoad: function() {
+    $('#js-connection-connection-type-input').chosen({width: '100%', disable_search: true}).change(() => {
+      const value = $('#js-connection-connection-type-input').val();
+      domEvents.dispatch($('#js-connection-connection-type-input')[0], {type: 'changed', data: {value}});
+    });
+    $('#js-default-user-select').chosen({width: '80%'}).change(() => {
+      const value = $('#js-default-user-select').val();
+      domEvents.dispatch($('#js-default-user-select')[0], {type: 'changed', data: {value}});
+    });
+    $('#js-default-group-admin-user-select').chosen({width: '80%'}).change(() => {
+      const value = $('#js-default-group-admin-user-select').val();
+      domEvents.dispatch($('#js-default-group-admin-user-select')[0], {type: 'changed', data: {value}});
+    });
   },
 
   /**
    * Init section
-   * @param {UsersDirectorySettings} usersDirectorySettings
    * @private
    */
-  _initFormGeneralSwitch: function(usersDirectorySettings) {
-    const enabled = usersDirectorySettings.isEnabled();
+  _initFormGeneralSwitch: function() {
+    const disabled = !this.options.edit;
+
+    const settings = this.options.usersDirectorySettings;
+    const enabled = settings.isEnabled();
     this.addElement(
       new ToggleButtonComponent('#js-ldap-toggle-button', {
         modelReference: 'UsersDirectorySettings.enabled',
         label: null,
-        state: {disabled: true},
+        state: {disabled},
         validate: false,
         value: enabled
       }).start()
@@ -99,56 +144,58 @@ const UsersDirectorySettingsForm = Form.extend('passbolt.form.administration.use
 
   /**
    * Init section
-   * @param {UsersDirectorySettings} usersDirectorySettings
    * @private
    */
   _initFormCredantialsSection: function() {
+    const disabled = !this.options.edit;
+
     this.addElement(
       new RadioComponent('#js-directory-type-radio', {
-        modelReference: 'UsersDirectorySettings.directory',
-        state: {disabled: true},
+        modelReference: 'UsersDirectorySettings.directory_type',
+        state: {disabled},
         availableValues: {
           'ad': __('Active Directory'),
-          'ldap': ('Open Ldap')
+          'openldap': ('Open Ldap')
         },
         value: 'ad'
       }).start()
     );
 
     this.addElement(
-      new TextboxComponent('#js-domain-input', {
-        modelReference: 'UsersDirectorySettings.domain',
-        state: {disabled: true}
+      new TextboxComponent('#js-domain-name-input', {
+        modelReference: 'UsersDirectorySettings.domain_name',
+        state: {disabled}
       }).start(),
-      new FeedbackComponent('#js-domain-input-feedback', {}).start()
+      new FeedbackComponent('#js-domain-name-input-feedback', {}).start()
     );
 
     this.addElement(
-      new DropdownComponent('#js-connection-protocol-input', {
-        modelReference: 'UsersDirectorySettings.protocol',
-        state: {disabled: true},
+      new DropdownComponent('#js-connection-connection-type-input', {
+        modelReference: 'UsersDirectorySettings.connection_type',
+        state: {disabled},
         availableValues: {
-          'ldap': 'ldap://',
-          'ldap_ssl': 'ldaps:// (ssl)',
-          'ldap_tls': 'ldaps:// (tls)'
+          'plain': 'ldap://',
+          'ssl': 'ldaps:// (ssl)',
+          'tls': 'ldaps:// (tls)'
         },
-        value: 'ldap_ssl',
+        value: 'plain',
         emptyValue: false
       }).start()
     );
 
     this.addElement(
-      new TextboxComponent('#js-host-input', {
-        modelReference: 'UsersDirectorySettings.host',
-        state: {disabled: true}
+      new TextboxComponent('#js-server-input', {
+        modelReference: 'UsersDirectorySettings.server',
+        state: {disabled}
       }).start(),
-      new FeedbackComponent('#js-host-input-feedback', {}).start()
+      new FeedbackComponent('#js-server-input-feedback', {}).start()
     );
 
     this.addElement(
       new TextboxComponent('#js-port-input', {
         modelReference: 'UsersDirectorySettings.port',
-        state: {disabled: true}
+        state: {disabled},
+        value: 389
       }).start(),
       new FeedbackComponent('#js-port-input-feedback', {}).start()
     );
@@ -156,7 +203,7 @@ const UsersDirectorySettingsForm = Form.extend('passbolt.form.administration.use
     this.addElement(
       new TextboxComponent('#js-username-input', {
         modelReference: 'UsersDirectorySettings.username',
-        state: {disabled: true}
+        state: {disabled}
       }).start(),
       new FeedbackComponent('#js-username-input-feedback', {}).start()
     );
@@ -164,7 +211,7 @@ const UsersDirectorySettingsForm = Form.extend('passbolt.form.administration.use
     this.addElement(
       new TextboxComponent('#js-password-input', {
         modelReference: 'UsersDirectorySettings.password',
-        state: {disabled: true}
+        state: {disabled}
       }).start(),
       new FeedbackComponent('#js-password-input-feedback', {}).start()
     );
@@ -172,7 +219,7 @@ const UsersDirectorySettingsForm = Form.extend('passbolt.form.administration.use
     this.addElement(
       new TextboxComponent('#js-base-dn-input', {
         modelReference: 'UsersDirectorySettings.base_dn',
-        state: {disabled: true}
+        state: {disabled}
       }).start(),
       new FeedbackComponent('#js-base-dn-input-feedback', {}).start()
     );
@@ -180,14 +227,15 @@ const UsersDirectorySettingsForm = Form.extend('passbolt.form.administration.use
 
   /**
    * Init section
-   * @param {UsersDirectorySettings} usersDirectorySettings
    * @private
    */
   _initFormDirectorySection: function() {
+    const disabled = !this.options.edit;
+
     this.addElement(
       new TextboxComponent('#js-group-path-input', {
         modelReference: 'UsersDirectorySettings.group_path',
-        state: {disabled: true}
+        state: {disabled}
       }).start(),
       new FeedbackComponent('#js-group-path-input-feedback', {}).start()
     );
@@ -195,7 +243,7 @@ const UsersDirectorySettingsForm = Form.extend('passbolt.form.administration.use
     this.addElement(
       new TextboxComponent('#js-user-path-input', {
         modelReference: 'UsersDirectorySettings.user_path',
-        state: {disabled: true}
+        state: {disabled}
       }).start(),
       new FeedbackComponent('#js-user-path-input-feedback', {}).start()
     );
@@ -203,7 +251,7 @@ const UsersDirectorySettingsForm = Form.extend('passbolt.form.administration.use
     this.addElement(
       new TextboxComponent('#js-group-object-class-input', {
         modelReference: 'UsersDirectorySettings.group_object_class',
-        state: {disabled: true}
+        state: {disabled}
       }).start(),
       new FeedbackComponent('#js-group-object-class-input-feedback', {}).start()
     );
@@ -211,108 +259,81 @@ const UsersDirectorySettingsForm = Form.extend('passbolt.form.administration.use
     this.addElement(
       new TextboxComponent('#js-user-object-class-input', {
         modelReference: 'UsersDirectorySettings.user_object_class',
-        state: {disabled: true}
+        state: {disabled}
       }).start(),
       new FeedbackComponent('#js-user-object-class-input-feedback', {}).start()
     );
 
     this.addElement(
-      new DropdownComponent('#js-default-admin-select', {
-        modelReference: 'UsersDirectorySettings.default_admin',
-        state: {disabled: true},
-        availableValues: []
+      new DropdownComponent('#js-default-user-select', {
+        modelReference: 'UsersDirectorySettings.default_user',
+        state: {disabled},
+        availableValues: this._admins
       }).start(),
-      new FeedbackComponent('#js-default-admin-select-feedback', {}).start()
+      new FeedbackComponent('#js-default-user-select-feedback', {}).start()
     );
 
     this.addElement(
-      new DropdownComponent('#js-default-group-admin-select', {
-        modelReference: 'UsersDirectorySettings.default_group_admin',
-        state: {disabled: true},
-        availableValues: []
+      new DropdownComponent('#js-default-group-admin-user-select', {
+        modelReference: 'UsersDirectorySettings.default_group_admin_user',
+        state: {disabled},
+        availableValues: this._users
       }).start(),
-      new FeedbackComponent('#js-default-group-admin-select-feedback', {}).start()
+      new FeedbackComponent('#js-default-group-admin-user-select-feedback', {}).start()
     );
   },
 
   /**
    * Init section
-   * @param {UsersDirectorySettings} usersDirectorySettings
    * @private
    */
   _initFormSynchronisationSection: function() {
+    const disabled = !this.options.edit;
+
     this.addElement(
-      new ToggleButtonComponent('#js-sync-create-users-toggle-button', {
+      new ToggleButtonComponent('#js-sync-users-create-toggle-button', {
         label: __('Create users'),
-        modelReference: 'UsersDirectorySettings.sync_create_users',
-        state: {disabled: true},
+        modelReference: 'UsersDirectorySettings.sync_users_create',
+        state: {disabled},
         value: true
       }).start()
     );
 
     this.addElement(
-      new ToggleButtonComponent('#js-sync-delete-users-toggle-button', {
+      new ToggleButtonComponent('#js-sync-users-delete-toggle-button', {
         label: __('Delete users'),
-        modelReference: 'UsersDirectorySettings.sync_delete_users',
-        state: {disabled: true},
+        modelReference: 'UsersDirectorySettings.sync_users_delete',
+        state: {disabled},
         value: true
       }).start()
     );
 
     this.addElement(
-      new ToggleButtonComponent('#js-sync-create-groups-toggle-button', {
+      new ToggleButtonComponent('#js-sync-groups-create-toggle-button', {
         label: __('Create users'),
-        modelReference: 'UsersDirectorySettings.sync_create_groups',
-        state: {disabled: true},
+        modelReference: 'UsersDirectorySettings.sync_groups_create',
+        state: {disabled},
         value: true
       }).start()
     );
 
     this.addElement(
-      new ToggleButtonComponent('#js-sync-delete-groups-toggle-button', {
+      new ToggleButtonComponent('#js-sync-groups-delete-toggle-button', {
         label: __('Delete users'),
-        modelReference: 'UsersDirectorySettings.sync_delete_groups',
-        state: {disabled: true},
+        modelReference: 'UsersDirectorySettings.sync_groups_delete',
+        state: {disabled},
         value: true
       }).start()
     );
 
     this.addElement(
-      new ToggleButtonComponent('#js-sync-update-groups-toggle-button', {
+      new ToggleButtonComponent('#js-sync-groups-update-toggle-button', {
         label: __('Create users'),
-        modelReference: 'UsersDirectorySettings.sync_update_groups',
-        state: {disabled: true},
+        modelReference: 'UsersDirectorySettings.sync_groups_update',
+        state: {disabled},
         value: true
       }).start()
     );
-  },
-
-  /**
-   * Init the edit form.
-   */
-  _initEditForm: function() {
-    const edit = this.options.edit;
-    if (!edit) {
-      return;
-    }
-
-    const findOptions = {
-      contain: {profile: 1},
-      filter: {'is-admin': 1}
-    };
-    return User.findAll(findOptions)
-      .then(users => {
-        const availableAdmins = users.reduce((carry, user) => {
-          carry[user.id] = `${user.profile.fullName()} (${user.username})`;
-          return carry;
-        }, {});
-
-        this.elements['js-default-admin-select'].setAvailableValues(availableAdmins);
-        this.elements['js-default-group-admin-select'].setAvailableValues(availableAdmins);
-        for (const i in this.elements) {
-          this.elements[i].state.disabled = false;
-        }
-      });
   },
 
   /**