+++ app/model/map/users_directory_settings.js
@@ -9,7 +9,7 @@
  * @copyright     Copyright (c) Passbolt SARL (https://www.passbolt.com)
  * @license       https://opensource.org/licenses/AGPL-3.0 AGPL License
  * @link          https://www.passbolt.com Passbolt(tm)
- * @since         2.5.0
+ * @since         2.6.0
  */
 import Ajax from 'app/net/ajax';
 import connect from 'can-connect';
@@ -30,24 +30,25 @@ const UsersDirectorySettings = DefineMap.extend('passbolt.model.UsersDirectorySe
     type: 'string',
     value: () => uuid()
   },
-  directory: 'string',
-  domain: 'string',
-  protocol: 'string',
-  host: 'string',
-  port: 'integer',
+  directory_type: 'string',
+  domain_name: 'string',
+  connection_type: 'string',
+  server: 'string',
+  port: 'number',
   username: 'string',
+  password: 'string',
   base_dn: 'string',
   group_path: 'string',
   user_path: 'string',
   group_object_class: 'string',
   user_object_class: 'string',
-  default_admin: 'string',
-  default_group_admin: 'string',
-  sync_create_users: 'boolean',
-  sync_delete_users: 'boolean',
-  sync_create_groups: 'boolean',
-  sync_delete_groups: 'boolean',
-  sync_update_groups: 'boolean'
+  default_user: 'string',
+  default_group_admin_user: 'string',
+  sync_users_create: 'boolean',
+  sync_users_delete: 'boolean',
+  sync_groups_create: 'boolean',
+  sync_groups_delete: 'boolean',
+  sync_groups_update: 'boolean'
 });
 DefineMap.setReference('UsersDirectorySettings', UsersDirectorySettings);
 
@@ -56,11 +57,11 @@ DefineMap.setReference('UsersDirectorySettings', UsersDirectorySettings);
  * Keep these rules in sync with the passbolt API.
  */
 UsersDirectorySettings.validationRules = {
-  domain: [
-    {rule: 'required', message: __('A domain is required.')},
-    {rule: 'utf8', message: __('The domain should be a valid utf8 string.')}
+  domain_name: [
+    {rule: 'required', message: __('A domain name is required.')},
+    {rule: 'utf8', message: __('The domain name should be a valid utf8 string.')}
   ],
-  host: [
+  server: [
     {rule: 'required', message: __('A host is required.')},
     {rule: 'utf8', message: __('The host should be a valid utf8 string.')}
   ],
@@ -81,46 +82,32 @@ UsersDirectorySettings.validationRules = {
     {rule: 'utf8', message: __('The base DN should be a valid utf8 string.')}
   ],
   group_path: [
-    {rule: 'required', message: __('A group path is required.')},
     {rule: 'utf8', message: __('The group path should be a valid utf8 string.')}
   ],
   user_path: [
-    {rule: 'required', message: __('A user path is required.')},
     {rule: 'utf8', message: __('The user path should be a valid utf8 string.')}
   ],
   group_object_class: [
-    {rule: 'required', message: __('A group object class is required.')},
     {rule: 'utf8', message: __('The group object class should be a valid utf8 string.')}
   ],
   user_object_class: [
-    {rule: 'required', message: __('A user object class is required.')},
     {rule: 'utf8', message: __('The user object class should be a valid utf8 string.')}
   ],
-  default_admin: [
+  default_user: [
     {rule: 'required', message: __('A default admin is required.')},
     {rule: 'uuid', message: __('The default admin should be a valid uuid.')}
   ],
-  default_group_admin: [
+  default_group_admin_user: [
     {rule: 'required', message: __('A default group admin is required.')},
-    {rule: 'utf8', message: __('The default group admin should be a valid uuid.')}
+    {rule: 'uuid', message: __('The default group admin should be a valid uuid.')}
   ]
-};
-
-/**
- * Disable the users directory
- */
-UsersDirectorySettings.prototype.disable = function() {
-  return Ajax.request({
-    url: 'settings/ldap/disable.json?api-version=v2',
-    type: 'PUT'
-  });
 };
 
 /**
  * Is a users directory enabled.
  */
 UsersDirectorySettings.prototype.isEnabled = function() {
-  if (this.domain) {
+  if (this.domain_name) {
     return true;
   }
 };
@@ -130,45 +117,25 @@ UsersDirectorySettings.connection = connect([connectParse, connectDataUrl, conne
   url: {
     resource: '/',
     getData: function() {
-      return new Promise(resolve => {
-        setTimeout(() => {
-          resolve({
-            directory: 'ldap',
-            domain: 'my domain',
-            protocol: 'ldap',
-            host: 'my host',
-            port: 999,
-            username: 'my username',
-            password: 'my password',
-            base_dn: 'my base_dn',
-            group_path: 'my group_path',
-            user_path: 'my user_path',
-            group_object_class: 'my group_object_class',
-            user_object_class: 'my user_object_class',
-            default_admin: 'd57c10f5-639d-5160-9c81-8a0c6c4ec856',
-            default_group_admin: 'd57c10f5-639d-5160-9c81-8a0c6c4ec856',
-            sync_create_users: true,
-            sync_delete_users: false,
-            sync_create_groups: true,
-            sync_delete_groups: false,
-            sync_update_groups: true
-          });
-        }, 1000);
+      const params = {};
+      params['api-version'] = 'v2';
+      return Ajax.request({
+        url: 'directorysync/settings.json',
+        type: 'GET',
+        params: params
       });
-      /*
-       *params['api-version'] = 'v2';
-       *return Ajax.request({
-       *  url: 'settings/ldap.json',
-       *  type: 'GET',
-       *  params: params
-       *});
-       */
     },
     updateData: function(params) {
       return Ajax.request({
-        url: 'settings/users_directory.json?api-version=v2',
+        url: 'directorysync/settings.json?api-version=v2',
         type: 'PUT',
         params: params
+      });
+    },
+    destroyData: function() {
+      return Ajax.request({
+        url: 'directorysync/settings.json?api-version=v2',
+        type: 'DELETE'
       });
     }
   }